move timers to frp, improve docs and variable names
