+++ lib/rust/ensogl/core/src/lib.rs
@@ -52,7 +52,6 @@ pub mod debug;
 pub mod display;
 pub mod gui;
 pub mod system;
-pub mod timer;
 
 pub use animation::Animation;
 pub use animation::DEPRECATED_Animation;
