+++ lib/rust/frp/src/io.rs
@@ -8,5 +8,6 @@
 pub mod js;
 pub mod keyboard;
 pub mod mouse;
+pub mod timer;
 
 pub use mouse::Mouse;
