+++ lib/rust/frp/src/io/timer.rs
@@ -1,9 +1,9 @@
 //! Browser timer handlers wrapped in FRP API.
 
-mod delayed_repeats;
+mod delayed_interval;
 mod interval;
 mod timeout;
 
-pub use delayed_repeats::*;
+pub use delayed_interval::*;
 pub use interval::*;
 pub use timeout::*;
